pyproject.toml
[tool.poetry]
name = "samberichecks"
version = "0.1.0"
description = ""
authors = ["onehandedpirate <plagueismkii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lxml = "^4.9.3"
beautifulsoup4 = "^4.12.2"
aiogram = "^3.1.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
